Solved chicken and egg problem by putting ng-show on .delayed-entrance themselves
Problem: need to wait for elements to be rendered to know which ones to
animate (via .delayed-entrance); but then there is a flicker. If we
don't wait long enough for the render, the .delayed-entrance might be
outdated and the elements will play the wrong animation. Solution: by
putting ng-show and .delayed-entrance on each element, they individually
start their own animation when they are ready, "independently" of the
others (which will play at the same time).